Added discounter and order

+++ checkout-kata/Interfaces/ICheckout.cs
@@ -4,6 +4,6 @@
     {
         void Scan(string sku);
 
-        decimal GetTotal();
+        double GetTotal();
     }
 }
