+++ checkout-kata/Models/Product.cs
@@ -3,6 +3,6 @@
     public class Product
     {
         public string Sku { get; set; }
-        public decimal UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
     }
 }
